+++ tests/autoarm/test_example_configs.py
@@ -4,7 +4,6 @@ from typing import Any
 
 import pytest
 from homeassistant import config as hass_config
-from homeassistant.components.calendar import CalendarEntity
 from homeassistant.config import (
     load_yaml_config_file,
 )
@@ -12,6 +11,7 @@ from homeassistant.const import SERVICE_RELOAD
 from homeassistant.core import HomeAssistant
 from homeassistant.setup import async_setup_component
 
+from conftest import MockAction
 from custom_components.autoarm.const import DOMAIN
 
 EXAMPLES_ROOT = Path("examples")
@@ -23,10 +23,11 @@ examples: list[Path] = [p for p in EXAMPLES_ROOT.iterdir() if p.is_file() and p.
 @pytest.mark.parametrize(argnames="reload", argvalues=[True, False], ids=["reload", "no-reload"])
 async def test_examples(
     hass: HomeAssistant,
-    config_name: str,
+    config_name: Path,
     reload: bool,
-    local_calendar: CalendarEntity,  # noqa: ARG001
+    test_config_calendars: None,  # noqa: ARG001
     monkeypatch: pytest.MonkeyPatch,
+    mock_notify: MockAction,  # noqa: ARG001
 ) -> None:
     config: dict[Any, Any] = await hass.async_add_executor_job(load_yaml_config_file, str(config_name))
 
@@ -44,8 +45,15 @@ async def test_examples(
     assert hass.states.get("sensor.autoarm_failures").state == "0"  # type: ignore
 
     config = await hass.services.async_call("autoarm", "enquire_configuration", None, blocking=True, return_response=True)
-    assert config["notify"]["common"]["service"] == "notify.send_message"
-    assert config["notify"]["quiet"]["source"] == ["alarm_panel", "button", "calendar", "sunrise", "sunset"]
+    if config_name.name == "typical.yaml":
+        assert config["notify"]["quiet"]["source"] == ["alarm_panel", "button", "sunrise", "sunset"]
+        assert config["notify"]["normal"]["source"] == ["calendar"]
+        assert config["notify"]["common"]["service"] == "notify.supernotify"
+        assert config["notify"]["common"]["supernotify"]
+    elif config_name.name == "minimal.yaml":
+        assert config["notify"]["quiet"]["source"] == ["alarm_panel", "button", "calendar", "sunrise", "sunset"]
+        assert config["notify"]["common"]["service"] == "notify.send_message"
+        assert not config["notify"]["common"]["supernotify"]
 
     if reload:
         config_path = pathlib.Path(__file__).parent.joinpath("fixtures", "empty_config.yaml").absolute()